Split file into two components
